fix: move orgDetailsInclude to a leaf module — the route was 500ing
Sharing the include from lib/data/org-details-server pulled requireOrgAccess,
and therefore the whole auth-helpers graph, into
GET /api/organizations/[orgId]. That route then returned 500 with an empty body
on every request, which is why the org dashboard still rendered only a skeleton:
the seed had nothing to seed and the client fetch failed too.

Confirmed by A/B rather than inspection, as the same OWNER session against both
previews:
  deploy-preview-1108 (no route change): 200, 3/3, real org payload
  deploy-preview-1107 (with it):         500, 3/3, empty body

The route only ever needed the SHAPE, so the shape now lives in
lib/data/org-details-include.ts with zero imports. The heavy read keeps its auth
dependency and imports the include like everyone else, so route and seed still
cannot drift.
